+++ src/test/java/im/aop/loggers/messageinterpolation/ObjectToStringStrategyTests.java
@@ -1,8 +1,11 @@
 package im.aop.loggers.messageinterpolation;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
 
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 /**
  * Tests for {@link ObjectToStringStrategy}.
@@ -56,6 +59,27 @@ class ObjectToStringStrategyTests {
   @Test
   void toString_givenObject() {
     final ObjectToStringStrategy toStringStrategy = new ObjectToStringStrategy();
+    assertThat(toStringStrategy.toString(new TestObject("foo"))).isEqualTo("[value=foo]");
+  }
+
+  @Test
+  void toString_withReflectionToStringStrategy() {
+    final ReflectionToStringStrategy reflectionToStringStrategy =
+        Mockito.mock(ReflectionToStringStrategy.class);
+    when(reflectionToStringStrategy.supports(any())).thenReturn(true);
+    when(reflectionToStringStrategy.toString(any())).thenReturn("x");
+
+    final ObjectToStringStrategy toStringStrategy = new ObjectToStringStrategy();
+    toStringStrategy.setReflectionToStringStrategy(reflectionToStringStrategy);
+
+    assertThat(toStringStrategy.toString(new TestObject("foo"))).isEqualTo("x");
+  }
+
+  @Test
+  void toString_withoutReflectionToStringStrategy() {
+    final ObjectToStringStrategy toStringStrategy = new ObjectToStringStrategy();
+    toStringStrategy.setReflectionToStringStrategy(null);
+
     assertThat(toStringStrategy.toString(new TestObject("foo"))).isEqualTo("[value=foo]");
   }
 }